+++ crypto/stark/benches/prover_benchmark.rs
@@ -3,7 +3,7 @@ use crypto::fiat_shamir::default_transcript::DefaultTranscript;
 use math::field::element::FieldElement;
 use math::field::fields::fft_friendly::extensions_goldilocks::Degree3GoldilocksExtensionField;
 use math::field::fields::fft_friendly::u64_goldilocks::GoldilocksField;
-use stark::examples::fibonacci_multi_column::{
+use stark::test_utils::examples::fibonacci_multi_column::{
     FibonacciMultiColumnAIR, FibonacciMultiColumnPublicInputs, compute_trace, create_public_inputs,
 };
 use stark::proof::options::ProofOptions;
