+++ crypto/stark/src/test_utils/examples/fibonacci_2_cols_shifted.rs
