+++ crypto/stark/src/test_utils/examples/fibonacci_2_columns.rs
