+++ crypto/stark/src/test_utils/examples/read_only_memory_logup.rs
