+++ crypto/stark/src/test_utils/examples/simple_fibonacci.rs
