+++ crypto/stark/src/test_utils/examples/simple_periodic_cols.rs
