+++ crypto/stark/src/test_utils/mod.rs
@@ -1,0 +1,4 @@
+//! Test utilities and example AIRs for testing and benchmarking
+
+#[cfg(any(test, feature = "test-utils"))]
+pub mod examples;
